Improve @ColumnTransformer @Where, Added @MappingEntity(alias="") (#926)
* Improve @ColumnTransformer @Where @MappedEntity(alias)

* Add docs

+++ data-jdbc/src/test/groovy/io/micronaut/data/jdbc/h2/H2EmbeddedIdSpec.groovy
@@ -77,6 +77,22 @@ class H2EmbeddedIdSpec extends Specification {
         then:"all is correct"
         all.size() == 2
 
+        when:"Find by country"
+        def foundByCountry = repository.findByShipmentIdCountry("g")
+
+        then:
+        foundByCountry.field == "test4"
+        foundByCountry.shipmentId.country == "g"
+        foundByCountry.shipmentId.city == "h"
+
+        when:"Find by country and city"
+        def foundByCountryAndCIty = repository.findByShipmentIdCountryAndShipmentIdCity("g", "h")
+
+        then:
+        foundByCountryAndCIty.field == "test4"
+        foundByCountryAndCIty.shipmentId.country == "g"
+        foundByCountryAndCIty.shipmentId.city == "h"
+
         when:"deleteAll is used with an iterable"
         repository.deleteAll([all.first()])
